pyproject.toml
[project]
name = "ibf-pipeline-core"
version = "0.0.1"
authors = [
  {name="Jacopo Margutti", email="jmargutti@redcross.nl" },
]
description = "IBF pipelines' core library"
readme = "README.md"
requires-python = ">=3.13,<4"
license = "AGPL-3.0"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/rodekruis/IBF-pipeline-core"
Issues = "https://github.com/rodekruis/IBF-pipeline-core/issues"

[tool.poetry]
packages = [
    { include = "core", from = "ibfpipeline" },
    { include = "drought", from = "ibfpipeline" },
    { include = "riverflood", from = "ibfpipeline" },
]

[tool.poetry.dependencies]
python = ">=3.13,<4"
azure-identity = "*"
azure-cosmos = "*"
azure-keyvault-secrets = "*"
azure-storage-blob = "*"
beautifulsoup4 = "*"
black = "*"
cdsapi = "*"
certifi = "*"
cfgrib = "*"
fiona = ">=1.10b2"
GeoAlchemy2 = "*"
geopandas = "*"
netCDF4 = "*"
pandas = "*"
PyYAML = "*"
python-dotenv = "*"
rasterio = "*"
rasterstats = ">=0.20.0"
rioxarray = "*"
scipy = "*"
SQLAlchemy = "*"
tqdm = "*"
xarray = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"